feat: Modular build system
This commit adds support for a modular build system. What is that?

What if you only care about non-root variant of both apps?
What if you just want to compile the root variant of YT? or non-root?

While using the CLI manually, it allows the user to choose which app they want
to build. You can choose to compile root or non-root variant of YT or YTM etc
However, with the script you would have to wait for both builds to compile.
Not anymore though.

You would have to wait for both of the build to compile. It was a waste of time
and resources. Now, not only will your build times be faster but you will
have control over particular app & its variant.